+++ src/planet.cpp
@@ -1,0 +1,5 @@
+#include "planet.h"
+
+Planet::Planet(float radius, float mass, std::string name) : CelestialObject(radius, mass, name){
+
+}