+++ src/spaceship.cpp
@@ -1,0 +1,5 @@
+#include "spaceship.h"
+
+Spaceship::Spaceship(std::string name) : CelestialObject(10.0f, 0, name){
+
+}